+++ llvm-hs-pure/src/LLVM/AST/Operand.hs
@@ -214,7 +214,7 @@ data DIScope
   deriving (Eq, Ord, Read, Show, Typeable, Data, Generic)
 
 data DIModule = Module
-  { scope :: MDRef DIScope
+  { scope :: Maybe (MDRef DIScope)
   , name :: ShortByteString
   , configurationMacros :: ShortByteString
   , includePath :: ShortByteString
@@ -223,7 +223,7 @@ data DIModule = Module
 
 data DINamespace = Namespace
   { name :: ShortByteString
-  , scope :: MDRef DIScope
+  , scope :: Maybe (MDRef DIScope)
   , exportSymbols :: Bool
   } deriving (Eq, Ord, Read, Show, Typeable, Data, Generic)
 
@@ -308,18 +308,18 @@ data DIType
 
 -- | <https://llvm.org/docs/LangRef.html#dibasictype>
 data DIBasicType = BasicType
-  { typeName :: ShortByteString
+  { name :: ShortByteString
   , sizeInBits :: Word64
   , alignInBits :: Word32
-  , typeEncoding :: Maybe Encoding
-  , typeTag :: BasicTypeTag
+  , encoding :: Maybe Encoding
+  , tag :: BasicTypeTag
   } deriving (Eq, Ord, Read, Show, Typeable, Data, Generic)
 
 -- | <https://llvm.org/docs/LangRef.html#disubroutinetype>
 data DISubroutineType = SubroutineType
-  { typeFlags :: [DIFlag]
-  , typeCC :: Word8
-  , typeTypeArray :: [Maybe (MDRef DIType)]
+  { flags :: [DIFlag]
+  , cc :: Word8
+  , typeArray :: [Maybe (MDRef DIType)]
   -- ^ The first element is the return type, the following are the
   -- operand types. `Nothing` corresponds to @void@.
   } deriving (Eq, Ord, Read, Show, Typeable, Data, Generic)
@@ -342,17 +342,17 @@ data DerivedTypeTag
 -- | <https://llvm.org/docs/LangRef.html#diderivedtype>
 data DIDerivedType =
   DerivedType
-    { derivedTag :: DerivedTypeTag
-    , derivedName :: ShortByteString
-    , derivedFile :: Maybe (MDRef DIFile)
-    , derivedLine :: Word32
-    , derivedScope :: Maybe (MDRef DIScope)
-    , derivedBaseType :: MDRef DIType
+    { tag :: DerivedTypeTag
+    , name :: ShortByteString
+    , file :: Maybe (MDRef DIFile)
+    , line :: Word32
+    , scope :: Maybe (MDRef DIScope)
+    , baseType :: MDRef DIType
     , sizeInBits :: Word64
     , alignInBits :: Word32
-    , derivedOffsetInBits :: Word64
-    , derivedAddressSpace :: Maybe Word32
-    , derivedFlags :: [DIFlag]
+    , offsetInBits :: Word64
+    , addressSpace :: Maybe Word32
+    , flags :: [DIFlag]
     } deriving (Eq, Ord, Read, Show, Typeable, Data, Generic)
 
 -- | <https://llvm.org/docs/LangRef.html#dicompositetype>