+++ llvm-hs/src/LLVM/Internal/Operand.hs
@@ -542,19 +542,19 @@ instance DecodeM DecodeAST A.DIBasicType (Ptr FFI.DIBasicType) where
 
 instance EncodeM EncodeAST A.DIBasicType (Ptr FFI.DIBasicType) where
   encodeM A.BasicType {..} = do
-    tag <- encodeM typeTag
-    typeName <- encodeM typeName
-    typeEncoding <- encodeM typeEncoding
+    tag <- encodeM tag
+    name <- encodeM name
+    encoding <- encodeM encoding
     Context c <- gets encodeStateContext
-    liftIO (FFI.getDIBasicType c tag typeName sizeInBits alignInBits typeEncoding)
+    liftIO (FFI.getDIBasicType c tag name sizeInBits alignInBits encoding)
 
 
 instance EncodeM EncodeAST A.DISubroutineType (Ptr FFI.DISubroutineType) where
   encodeM A.SubroutineType {..} = do
-    flags <- encodeM typeFlags
-    types <- encodeM typeTypeArray
+    flags <- encodeM flags
+    types <- encodeM typeArray
     Context c <- gets encodeStateContext
-    liftIO (FFI.getDISubroutineType c flags typeCC types)
+    liftIO (FFI.getDISubroutineType c flags cc types)
 
 instance DecodeM DecodeAST A.DISubroutineType (Ptr FFI.DISubroutineType) where
   decodeM p = do
@@ -562,23 +562,23 @@ instance DecodeM DecodeAST A.DISubroutineType (Ptr FFI.DISubroutineType) where
     cc <-  liftIO (FFI.getSubroutineCC p)
     arr <- decodeM =<< liftIO (FFI.getSubroutineTypeArray p)
     pure A.SubroutineType
-      { A.typeFlags = flags
-      , A.typeCC = cc
-      , A.typeTypeArray = arr
+      { A.flags = flags
+      , A.cc = cc
+      , A.typeArray = arr
       }
 
 instance EncodeM EncodeAST A.DIDerivedType (Ptr FFI.DIDerivedType) where
   encodeM A.DerivedType {..} = do
-    tag <- encodeM derivedTag
-    name <- encodeM derivedName
-    file <- encodeM derivedFile
-    line <- encodeM derivedLine
-    scope <- encodeM derivedScope
-    type' <- encodeM derivedBaseType
-    (addrSpace, addrSpacePresent) <- encodeM derivedAddressSpace
-    flags <- encodeM derivedFlags
+    tag <- encodeM tag
+    name <- encodeM name
+    file <- encodeM file
+    line <- encodeM line
+    scope <- encodeM scope
+    type' <- encodeM baseType
+    (addrSpace, addrSpacePresent) <- encodeM addressSpace
+    flags <- encodeM flags
     Context c <- gets encodeStateContext
-    FFI.upCast <$> liftIO (FFI.getDIDerivedType c tag name file line scope type' sizeInBits alignInBits derivedOffsetInBits addrSpace addrSpacePresent flags)
+    FFI.upCast <$> liftIO (FFI.getDIDerivedType c tag name file line scope type' sizeInBits alignInBits offsetInBits addrSpace addrSpacePresent flags)
 
 instance DecodeM DecodeAST A.DIDerivedType (Ptr FFI.DIDerivedType) where
   decodeM diTy = do
@@ -595,17 +595,17 @@ instance DecodeM DecodeAST A.DIDerivedType (Ptr FFI.DIDerivedType) where
     flags <- decodeM =<< liftIO (FFI.getTypeFlags diTy')
     addressSpace <- decodeOptional (FFI.getDerivedAddressSpace diTy')
     pure A.DerivedType
-      { A.derivedTag = tag
-      , A.derivedName = name
-      , A.derivedFile = file
-      , A.derivedLine = line
-      , A.derivedScope = scope
-      , A.derivedBaseType = ty
+      { A.tag = tag
+      , A.name = name
+      , A.file = file
+      , A.line = line
+      , A.scope = scope
+      , A.baseType = ty
       , A.sizeInBits = size
       , A.alignInBits = align
-      , A.derivedOffsetInBits = offset
-      , A.derivedAddressSpace = addressSpace
-      , A.derivedFlags = flags
+      , A.offsetInBits = offset
+      , A.addressSpace = addressSpace
+      , A.flags = flags
       }
 
 genCodingInstance [t|A.DerivedTypeTag|] ''FFI.DwTag